% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{dcor}
\alias{dcor}
\title{Distance Correlation}
\usage{
dcor(x,y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector}
}
\value{
Returns a numeric value: the distance correlation between x and y.
}
\description{
Distance Correlation
}
\examples{
\dontrun{

set.seed(1)
x < -rnorm(1000)
y < -x ^ 2

dcor(x, y) # dcor shows dependence between x and y
cor(x, y) # cor does not detect any depencence due to nonlinearity

}
}
